+++ flask_commands/project/app/controllers/main_controller.py
@@ -1,6 +1,5 @@
 from flask import render_template
 
 class MainController:
-    @staticmethod
-    def index() -> str:
+    def index(self) -> str:
         return render_template('mains/index.html')