+++ flask_commands/project/app/routes/mains/routes.py
@@ -3,4 +3,4 @@ from app.routes.mains import bp
 
 @bp.route('/', methods=['GET'])
 def index():
-    return MainController.index()
+    return MainController().index()